Move Content Type inside encryption
Previous versions of TLS have left the Content Type in the clear
without justifying the exposure of this piece of information.

This change moves the Content Type inside the encryption to protect
its confidentiality.

For backward compatibility during the initial handshake negotiation,
we need to keep the type field at its original location on the wire
for cleartext packets.  This means redefining the NULL_NULL case to
not simply be an pseudo-cipher that implements the identity
transformation, but to explicitly state it as sending the TLSPlaintext
struct directly over the wire.